Broke ListViewBase into 3 partials
